Introducing a general base form for entity alias configuration.
